Move ViewTracker to separate header file (from Kokkos_View.hpp) address issues from PR:    change typedefs to using    added explicit where appropriate    use assign(ViewType &) where appropriate    fixed friend clause    remove unecessary accessor class and reference friend members    add addition comments to clarify function.
