+++ core/src/impl/Kokkos_SharedAlloc.hpp
@@ -394,6 +394,14 @@ union SharedAllocationTracker {
     return *this;
   }
 
+  /*  The following functions (assign_direct and assign_force_disable)
+   *  are the result of deconstructing the
+   *  KOKKOS_IMPL_SHARED_ALLOCATION_CARRY_RECORD_BITS macro.  This
+   *  allows the caller to do the check for tracking enabled and managed
+   *  apart from the assignement of the record because the tracking
+   *  enabled / managed question may be important for other tasks as well
+   */
+
   /** \brief  Copy assignment without the carry bits logic
    *         This assumes that externally defined tracking is explicitly enabled
    */
@@ -404,9 +412,10 @@ union SharedAllocationTracker {
     KOKKOS_IMPL_SHARED_ALLOCATION_TRACKER_INCREMENT
   }
 
-  /** \brief  Copy assignment without the initial decrement
-   *         This assumes that current m_record_bits is DO_NOT_DEREF_FLAG
-   *         and externally defined tracking is explicitly disabled */
+  /** \brief  Copy assignment without the increment
+   *         we cannot assume that current record is unmanaged
+   *         but with externally defined tracking explicitly disabled
+   *         we can go straight to the do not deref flag     */
   KOKKOS_FORCEINLINE_FUNCTION
   void assign_force_disable(const SharedAllocationTracker& rhs) {
     KOKKOS_IMPL_SHARED_ALLOCATION_TRACKER_DECREMENT
